Bug fixes
Working through a ton of bugs. This version will at least run through
the sim. Fire round has been commented out of runSim() because the
functionality is off; when left in, all organisms are killed within
only a few rounds. It also seems that the organisms are not propagating
properly. Tree objects quickly take over but then appear stagnate,
without continuing to reproduce until the zone is full. This may be an
issue having to do with the getAdjZones() implementation and how
treeTurn(), grassTurn(), and deerTurn() use it.